+++ src/Sentry.Unity.Editor/ScriptableSentryUnityOptionsEditor.cs
@@ -66,6 +66,13 @@ namespace Sentry.Unity.Editor
             EditorGUI.DrawRect(EditorGUILayout.GetControlRect(false, 1), Color.gray);
             EditorGUILayout.Space();
 
+            EditorGUILayout.ObjectField("Options Configuration", options.OptionsConfiguration,
+                typeof(ScriptableOptionsConfiguration), false);
+
+            EditorGUILayout.Space();
+            EditorGUI.DrawRect(EditorGUILayout.GetControlRect(false, 1), Color.gray);
+            EditorGUILayout.Space();
+
             EditorGUILayout.LabelField("Debug", EditorStyles.boldLabel);
             EditorGUILayout.Toggle("Enable Debug Output", options.Debug);
             EditorGUILayout.Toggle("Only In Editor", options.DebugOnlyInEditor);